+++ docs/changelog.md
@@ -11,7 +11,7 @@ nav_order: 2
 ### Internals
 
 - Return `bytes` from `BaseAddress.to_knx()` instead of `tuple[int, int]`. This is used in `IndividualAddress` and `GroupAddress`.
-- Add `BaseAddress.from_knx()` to instantiate from `bytes`.
+- Add `BaseAddress.from_knx()` to instantiate from `bytes`, remove instantiation form `tuple[int, int]`.
 
 # 2.3.0 Routing security, DPTs and CEMI-Refactoring 2023-01-10